moved dist and sdes into seperate folders. Cleaned up type diff
