+++ src/agedi/diffusion/noisers/base.py
@@ -1,7 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import Type, Dict
 
-from .sdes import SDE, VE, VP
 from .distributions import Distribution
 from agedi.data import AtomsGraph
 
@@ -15,10 +13,6 @@ class Noiser(ABC, torch.nn.Module):
 
     Parameters
     ----------
-    sde_class: Type[SDE]
-        The class of the SDE to be used for the noising and denoising.
-    sde_kwargs: dict
-        The keyword arguments to be passed to the SDE class.
     distribution: Distribution
         The distribution to be used for the noising.
     prior: Distribution
@@ -31,23 +25,19 @@ class Noiser(ABC, torch.nn.Module):
     Noiser
 
     """
+
     _key: str
-    
+
     def __init__(
         self,
-        sde_class: Type[SDE],
-        sde_kwargs: Dict,
         distribution: Distribution,
         prior: Distribution,
-        loss_scaling: float=1.0,
+        loss_scaling: float = 1.0,
         **kwargs
     ):
-        """Initializes the Noiser.
-        
-        """
+        """Initializes the Noiser."""
         super().__init__(**kwargs)
-        self.sde = sde_class(**sde_kwargs)
-        
+
         self.distribution = distribution
         self.distribution.key = self.key
 
@@ -55,22 +45,18 @@ class Noiser(ABC, torch.nn.Module):
         self.prior.key = self.key
 
         self.loss_scaling = loss_scaling
-        
 
     @property
     def key(self) -> str:
-        """The key of the attribute to be noised and denoised.
-
-        """
+        """The key of the attribute to be noised and denoised."""
         return self._key
 
-    
     @abstractmethod
     def _noise(self, batch: AtomsGraph) -> AtomsGraph:
         """Noises the attribute of the atomistic structure.
 
         Must be implemented by the subclass.
-        
+
         Parameters
         ----------
         batch: AtomsGraph
@@ -80,7 +66,7 @@ class Noiser(ABC, torch.nn.Module):
         -------
         AtomsGraph
             The noised atomistic structure (or bach hereof).
-        
+
         """
         pass
 
@@ -89,7 +75,7 @@ class Noiser(ABC, torch.nn.Module):
         """Denoises the attribute of the atomistic structure.
 
         Must be implemented by the subclass.
-        
+
         Parameters
         ----------
         batch: AtomsGraph
@@ -102,7 +88,7 @@ class Noiser(ABC, torch.nn.Module):
         -------
         AtomsGraph
             The denoised atomistic structure (or bach hereof).
-        
+
         """
         pass
 
@@ -137,7 +123,7 @@ class Noiser(ABC, torch.nn.Module):
         -------
         AtomsGraph
             The noised atomistic structure (or bach hereof).
-        
+
         """
         return self._noise(batch)
 
@@ -150,12 +136,14 @@ class Noiser(ABC, torch.nn.Module):
             The atomistic structure (or batch hereof) to be denoised.
         delta_t: float
             The time step to be used for the denoising.
+        last: bool
+            If the denoising is the last step of the denoising.
 
         Returns
         -------
         AtomsGraph
             The denoised atomistic structure (or bach hereof).
-        
+
         """
         return self._denoise(batch, delta_t, last)
 
@@ -174,4 +162,3 @@ class Noiser(ABC, torch.nn.Module):
 
         """
         return self._loss(batch)
-
